Transferred the project to a multi-module structure
