--- app/src/main/java/com/kerencev/translator/data/dto/Translation.kt
@@ -1,10 +1,0 @@
-package com.kerencev.translator.data.dto
-
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Translation(
-    @field:SerializedName("text") val translation: String?
-) : Parcelable
