--- app/src/main/java/com/kerencev/translator/data/local/dao/HistoryDao.kt
@@ -1,17 +1,0 @@
-package com.kerencev.translator.data.local.dao
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.kerencev.translator.data.local.entity.HistoryEntity
-
-@Dao
-interface HistoryDao {
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insert(historyEntity: HistoryEntity)
-
-    @Query("SELECT * FROM history ORDER BY id DESC")
-    suspend fun getAll(): List<HistoryEntity>
-}
