--- app/src/main/java/com/kerencev/translator/data/remote/ApiService.kt
@@ -1,12 +1,0 @@
-package com.kerencev.translator.data.remote
-
-import com.kerencev.translator.data.dto.DataModel
-import kotlinx.coroutines.Deferred
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ApiService {
-
-    @GET("words/search")
-    fun searchAsync(@Query("search") wordToSearch: String): Deferred<List<DataModel>>
-}
