--- app/src/main/java/com/kerencev/translator/data/remote/BaseInterceptor.kt
@@ -1,18 +1,0 @@
-package com.kerencev.translator.data.remote
-
-import okhttp3.Interceptor
-import okhttp3.Response
-import java.io.IOException
-
-class BaseInterceptor private constructor() : Interceptor {
-
-    @Throws(IOException::class)
-    override fun intercept(chain: Interceptor.Chain): Response {
-        return chain.proceed(chain.request())
-    }
-
-    companion object {
-        val interceptor: BaseInterceptor
-            get() = BaseInterceptor()
-    }
-}
