+++ app/src/main/java/com/kerencev/translator/presentation/details/DetailsFragment.kt
@@ -1,18 +1,8 @@
 package com.kerencev.translator.presentation.details
 
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
+import com.kerencev.domain.model.DetailsModel
 
 interface DetailsFragment {
     fun getDataFromArgs(): DetailsModel?
     fun renderData(data: DetailsModel)
 }
-
-@Parcelize
-data class DetailsModel(
-    val id: String,
-    val word: String?,
-    val transcription: String?,
-    val translates: String,
-    val imageUrl: String?
-) : Parcelable
