+++ data/src/main/java/com/kerencev/data/dto/Translation.kt
@@ -1,0 +1,10 @@
+package com.kerencev.data.dto
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Translation(
+    @field:SerializedName("text") val translation: String?
+) : Parcelable
