+++ data/src/main/java/com/kerencev/data/local/dao/HistoryDao.kt
@@ -1,0 +1,17 @@
+package com.kerencev.data.local.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.kerencev.data.local.entity.HistoryEntity
+
+@Dao
+interface HistoryDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insert(historyEntity: HistoryEntity)
+
+    @Query("SELECT * FROM history ORDER BY id DESC")
+    suspend fun getAll(): List<HistoryEntity>
+}
