+++ data/src/main/java/com/kerencev/data/remote/ApiService.kt
@@ -1,0 +1,12 @@
+package com.kerencev.data.remote
+
+import com.kerencev.data.dto.DataModel
+import kotlinx.coroutines.Deferred
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ApiService {
+
+    @GET("words/search")
+    fun searchAsync(@Query("search") wordToSearch: String): Deferred<List<DataModel>>
+}
