+++ domain/src/main/java/com/kerencev/domain/Repository.kt
@@ -1,0 +1,8 @@
+package com.kerencev.domain
+
+import com.kerencev.data.dto.DataModel
+
+interface Repository<T : Any> {
+
+    suspend fun getData(word: String): List<DataModel>
+}
